Make bucket list non-optional
Refactor the bucket values, it is now all back in GaugeValue.
If buckets are not wanted then the user passes the empty list to
the function `addGauge`. The +Inf is then automatically removed.
So it is not possible to have only the +Inf bucket and no other
bucket. The counter for +Inf is also removed, we use the already
existing counter `count` instead for that value.

The metrics sum, count and bucket kind of belong together in
Prometheus, hence it is nicer to not have them spreaded over two
different Values.